refactor: Se hizo un refactor para un mejor manejo de excepciones y colocar bien la logica
Se cambiaron los siguientes detalles:
Se quito un Package que no se necesitaba en FibonacciExercise.IntegretedTest
Se cambiaron las prop a private en FibonacciTestIntegracion, también el Assert.IsTrue por Assert.AreEqual. Ahora todos los Assert.AreEqual tienen bien el orden de los parámetros.
En el FibonacciController quite la lógica, y cambie por un try catch, para que devolviera un BadRequest. Quite el codigo comentado.
En IFibonacciServicios, cambie el retorno por un long.
En FibonacciServicios, cambie que en casos de que el numero sea mayor a 90 o menor a 0 retorne un exception. Y que tanto CalcularFibonacci como RealizarFibonacci retorne un long, y en RealizarFibonacci espera un long.
En FibonacciTest (los test unitarios), separe los  test por categoria y utilive el DataRow para pasarles los parametros.
Se modificó la descripción del proyecto, del readme
Se cambiarón los mensajes, a constantes. Las respuestas estan en Respuestas, Constantes.

+++ src/FibonacciWebApi/Servicios/IFibonacciServicios.cs
@@ -7,6 +7,6 @@ namespace FibonacciWebApi.Servicios
 {
     public interface IFibonacciServicios
     {
-        int CalcularFibonacci(int i);
+        long CalcularFibonacci(int i);
     }
 }
